LBAP-0021 | API: Clients (Admin) | Кнопка «отработано» | Изменение статуса заявки
